Fix incorrect arguments to log functions.

+++ src/lib/socket/rrr_socket_client.c
@@ -1563,7 +1563,8 @@ static int __rrr_socket_client_collection_find_by_address_string_or_connect (
 				RRR_BUG("BUG: FD not set after connect callback in __rrr_socket_client_collection_find_by_address_string_or_connect\n");
 			}
 
-			RRR_DBG_7("client collection connect to '%s' suggestion %i/%i now pending\n", addr_string, i + 1, address_count);
+			RRR_DBG_7("client collection connect to '%s' suggestion %llu/%llu now pending\n",
+					addr_string, (long long unsigned int) i + 1, (long long unsigned int) address_count);
 
 			if ((ret = __rrr_socket_client_not_ready_fd_push (
 					client,
@@ -1578,7 +1579,8 @@ static int __rrr_socket_client_collection_find_by_address_string_or_connect (
 			tmp_fd = -1;
 		}
 		else {
-			RRR_DBG_7("client collection connect to '%s' suggestion %i/%i failed\n", addr_string, i + 1, address_count);
+			RRR_DBG_7("client collection connect to '%s' suggestion %llu/%llu failed\n",
+					addr_string, (long long unsigned int) i + 1, (long long unsigned int) address_count);
 		}
 	}
